+++ src/main/java/patcher/remote_api/entities/Patch.java
@@ -13,16 +13,16 @@ public class Patch {
     private String versionTo;
 
     @Getter @Setter
-    private Integer filesCount;
+    private Long filesCount;
 
     @Getter @Setter
-    private Integer totalSize;
+    private Long totalSize;
 
     public Patch(JSONObject patch) {
         this.versionFrom = patch.getString("version_from");
         this.versionTo = patch.getString("version_to");
-        this.filesCount = patch.getInt("files_count");
-        this.totalSize = patch.getInt("total_size");
+        this.filesCount = patch.getLong("files_count");
+        this.totalSize = patch.getLong("total_size");
     }
 
     public JSONObject toJSON() {
